Added documentation for a successful test and made dependencies not global necessary
Dependencies will now be called via their node_mdoules path when running scripts like npm run xxxxx

+++ scripts/win_install.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 
 echo '>> INSTALL [WINDOWS]'
+
+# Check if npm is installed
+if ! [ -x "$(command -v npm)" ]; then
+  echo 'Error: npm is not installed. (https://nodejs.org/en/download/current/)' >&2
+  exit 1
+fi
+
 cd ..
 # Remove old build
 rm -fr dist
